% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggrHiCExperiment-methods.R
\name{AggrHiCExperiment-methods}
\alias{AggrHiCExperiment-methods}
\alias{slices,AggrHiCExperiment,missing-method}
\alias{slices,AggrHiCExperiment,character-method}
\alias{slices,AggrHiCExperiment,numeric-method}
\alias{show,AggrHiCExperiment-method}
\title{\code{AggrHiCExperiment} methods}
\arguments{
\item{x}{A \code{AggrHiCExperiment} object.}

\item{name}{The name/index of slices to extract.}
}
\description{
AggrHiCExperiment methods.
}
