% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggrHiCExperiment-class.R
\docType{class}
\name{AggrHiCExperiment}
\alias{AggrHiCExperiment}
\alias{AggrHiCExperiment-class}
\alias{.AggrHiCExperimentFromCoolFile}
\alias{.AggrHiCExperimentFromHicFile}
\alias{.AggrHiCExperimentFromHicproFile}
\title{\code{AggrHiCExperiment} S4 class}
\usage{
AggrHiCExperiment(
  file,
  resolution = NULL,
  snippets,
  metadata = list(),
  topologicalFeatures = S4Vectors::SimpleList(loops =
    S4Vectors::Pairs(GenomicRanges::GRanges(), GenomicRanges::GRanges()), borders =
    GenomicRanges::GRanges(), compartments = GenomicRanges::GRanges(), viewpoints =
    GenomicRanges::GRanges()),
  pairsFile = NULL,
  BPPARAM = BiocParallel::SerialParam(progressbar = TRUE),
  ...
)

.AggrHiCExperimentFromCoolFile(
  file,
  resolution = NULL,
  snippets,
  metadata = list(),
  topologicalFeatures = S4Vectors::SimpleList(loops =
    S4Vectors::Pairs(GenomicRanges::GRanges(), GenomicRanges::GRanges()), borders =
    GenomicRanges::GRanges(), compartments = GenomicRanges::GRanges(), viewpoints =
    GenomicRanges::GRanges()),
  pairsFile = NULL,
  BPPARAM
)

.AggrHiCExperimentFromHicFile(
  file,
  resolution = NULL,
  snippets,
  metadata = list(),
  topologicalFeatures = S4Vectors::SimpleList(loops =
    S4Vectors::Pairs(GenomicRanges::GRanges(), GenomicRanges::GRanges()), borders =
    GenomicRanges::GRanges(), compartments = GenomicRanges::GRanges(), viewpoints =
    GenomicRanges::GRanges()),
  pairsFile = NULL,
  BPPARAM
)

.AggrHiCExperimentFromHicproFile(
  file,
  bed,
  snippets,
  topologicalFeatures = S4Vectors::SimpleList(loops =
    S4Vectors::Pairs(GenomicRanges::GRanges(), GenomicRanges::GRanges()), borders =
    GenomicRanges::GRanges(), compartments = GenomicRanges::GRanges(), viewpoints =
    GenomicRanges::GRanges()),
  pairsFile = NULL,
  BPPARAM
)
}
\description{
The \code{AggrHiCExperiment} extends \code{HiCExperiment} class
}
\section{Slots}{

\describe{
\item{\code{fileName}}{Path of Hi-C contact file}

\item{\code{resolutions}}{Resolutions available in the Hi-C contact file.}

\item{\code{resolution}}{Current resolution}

\item{\code{interactions}}{Genomic Interactions extracted from the Hi-C contact file}

\item{\code{scores}}{Available interaction scores.}

\item{\code{topologicalFeatures}}{Topological features associated with the dataset
(e.g. loops (\<Pairs\>), borders (\<GRanges\>),
viewpoints (\<GRanges\>), etc...)}

\item{\code{pairsFile}}{Path to the .pairs file associated with the Hi-C contact file}

\item{\code{metadata}}{metadata associated with the Hi-C contact file.}
}}

\examples{
library(rtracklayer)
devtools::load_all('.')
file = "/home/rsg/.cache/R/fourDNData/78685597e138_4DNFI9FVHJZQ.mcool"
file = "/data/20221214_HiContacts_compartments-TADs-loops/data/test.mcool"
resolution = 10000 
BPPARAM = BiocParallel::SerialParam(progressbar = TRUE)
coords_list = import('/home/rsg/Projects/20221214_HiContacts_compartments-TADs-loops/data/test_CTCF.bed.gz')
CTCF_bw <- import('/home/rsg/Projects/20221214_HiContacts_compartments-TADs-loops/data/test_CTCF.bigWig', as = 'NumericList', selection = BigWigSelection(resize(coords_list, fix = 'center', width = 250)))
coords_list = import('/home/rsg/Projects/20221214_HiContacts_compartments-TADs-loops/data/test_CTCF.bed.gz')
coords_list$scoreChIP <- mean(CTCF_bw)
snippets <- coords_list[coords_list$score >= quantile(coords_list$score , 0.75) & coords_list$scoreChIP >= quantile(coords_list$scoreChIP , 0.75)]
snippets <- GenomicRanges::resize(snippets, width = resolution*100, fix = 'center')
snippets <- snippets[order(snippets$scoreChIP, decreasing = TRUE)]
pairs <- Pairs(
    rep(snippets, each = length(snippets)),
    rep(snippets, length(snippets))
)
d <- distance(pairs)
pairs <- pairs[d > 200000 & d < 1000000 & !is.na(d)]
}
