% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{AllGenerics}
\alias{AllGenerics}
\title{Generic functions}
\arguments{
\item{x}{Passed to corresponding method}

\item{name}{Passed to corresponding method}

\item{value}{Passed to corresponding method}
}
\description{
Generics functions created in HiCExperiment package.
}
