% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin-methods}
\alias{bin-methods}
\alias{bin,GInteractions,character-method}
\alias{bin,GInteractions-method}
\title{HiCExperiment binning methods}
\usage{
\S4method{bin}{GInteractions}(x, resolution, seqinfo = NULL, pairsFile = NULL)
}
\description{
HiCExperiment binning methods
}
\examples{
pairsf <- HiContactsData::HiContactsData('yeast_wt', 'pairs.gz')
pf <- PairsFile(pairsf)

}
