% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{checks}
\alias{checks}
\alias{check_cool_file}
\alias{check_resolution}
\alias{check_cool_format}
\alias{is_mcool}
\alias{is_cool}
\alias{is_square}
\title{Checks functions}
\usage{
check_cool_file(path)

check_resolution(contacts, resolution)

check_cool_format(path, resolution, ...)

is_mcool(path)

is_cool(path)

is_square(pair)
}
\arguments{
\item{path}{Path of a (m)cool file}

\item{contacts}{A \code{HiCExperiment} object}

\item{resolution}{Resolution}

\item{...}{\code{HiCExperiment} object, arguments passed on by other functions}

\item{pair}{Pairs object with length of 1}
}
\value{
Logical
}
\description{
Useful functions to validate the nature/structure of (m)cool files or
\code{HiCExperiment} objects.
All these check functions should return a logical.
}
