% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContactsFile-methods.R
\name{ContactsFile-methods}
\alias{ContactsFile-methods}
\alias{pairsFile,ContactsFile-method}
\alias{resolution,ContactsFile-method}
\alias{metadata<-,ContactsFile-method}
\alias{metadata<-,ContactsFile,list-method}
\title{\code{ContactsFile} methods}
\arguments{
\item{object}{A \code{ContactsFile} object.}

\item{x}{A \code{ContactsFile} object.}
}
\description{
ContactsFile methods.
}
