% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoolFile-class.R
\name{CoolFile-class}
\alias{CoolFile-class}
\alias{McoolFile-class}
\alias{CoolFile}
\title{\code{CoolFile} S4 class}
\arguments{
\item{path}{String; path to a (m)cool file}

\item{resolution}{numeric; resolution to use with an mcool file}
}
\description{
The \code{CoolFile} class describes a \code{BiocFile} object, pointing to the location
of an (m)cool file and containing an additional \code{resolution} slot,
specifying at which resolution the associated mcool file should be parsed
when imported in R.
}
\section{Slots}{

\describe{
\item{\code{resolution}}{numeric value or NULL}
}}

\examples{
cool_path <- HiContactsData::HiContactsData('yeast_wt', 'cool')
CoolFile(cool_path)
cool_path <- HiContactsData::HiContactsData('yeast_wt', 'mcool')
cf <- CoolFile(cool_path, resolution = 4000)
}
