% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoolFile-methods.R
\name{CoolFile-methods}
\alias{CoolFile-methods}
\alias{show,CoolFile-method}
\alias{resolution,CoolFile-method}
\title{\code{CoolFile} methods}
\arguments{
\item{object}{A \code{CoolFile} object.}

\item{x}{A \code{CoolFile} object.}
}
\description{
CoolFile methods.
}
\examples{
cool_path <- HiContactsData::HiContactsData('yeast_wt', 'mcool')
cf <- CoolFile(cool_path, resolution = 4000)
cf
resolution(cf)
}
