% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoolFile-class.R, R/CoolFile-method.R
\docType{class}
\name{CoolFile-class}
\alias{CoolFile-class}
\alias{CoolFile}
\alias{resolution}
\alias{resolution,CoolFile-method}
\alias{show}
\alias{show,CoolFile-method}
\title{\code{CoolFile} S4 class and methods}
\usage{
CoolFile(path, resolution = NULL)

\S4method{resolution}{CoolFile}(x)

\S4method{show}{CoolFile}(object)
}
\arguments{
\item{object}{A \code{CoolFile} object.}
}
\description{
\code{CoolFile} S4 class and methods
}
\section{Slots}{

\describe{
\item{\code{resolution}}{numeric value or NULL}
}}

\examples{
cool_path <- HiContactsData::HiContactsData('yeast_wt', 'cool')
CoolFile(cool_path)
cool_path <- HiContactsData::HiContactsData('yeast_wt', 'mcool')
CoolFile(cool_path, resolution = 1000)
cool_path <- HiContactsData::HiContactsData('yeast_wt', 'cool')
cfile <- CoolFile(cool_path)
resolution(cfile)
mcool_path <- HiContactsData::HiContactsData('yeast_wt', 'mcool')
cfile <- CoolFile(mcool_path, resolution)
resolution(cfile)
show(contacts_yeast)
}
