% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiCExperiment-methods.R
\name{HiCExperiment-methods}
\alias{HiCExperiment-methods}
\alias{resolutions,HiCExperiment-method}
\alias{resolution,HiCExperiment-method}
\alias{focus,HiCExperiment-method}
\alias{focus<-,HiCExperiment-method}
\alias{focus<-,HiCExperiment,character-method}
\alias{zoom,HiCExperiment,numeric-method}
\alias{refocus,HiCExperiment,character-method}
\alias{scores,HiCExperiment-method}
\alias{scores,HiCExperiment,missing-method}
\alias{scores,HiCExperiment,character-method}
\alias{scores,HiCExperiment,numeric-method}
\alias{scores<-,HiCExperiment-method}
\alias{scores<-,HiCExperiment,character,numeric-method}
\alias{topologicalFeatures,HiCExperiment-method}
\alias{topologicalFeatures,HiCExperiment,missing-method}
\alias{topologicalFeatures,HiCExperiment,character-method}
\alias{topologicalFeatures,HiCExperiment,numeric-method}
\alias{topologicalFeatures<-,HiCExperiment-method}
\alias{topologicalFeatures<-,HiCExperiment,character,GRangesOrGInteractions-method}
\alias{pairsFile,HiCExperiment-method}
\alias{pairsFile<-,HiCExperiment-method}
\alias{pairsFile<-,HiCExperiment,character-method}
\alias{metadata<-,HiCExperiment-method}
\alias{metadata<-,HiCExperiment,list-method}
\alias{fileName,HiCExperiment-method}
\alias{interactions,HiCExperiment-method}
\alias{interactions<-,HiCExperiment-method}
\alias{length,HiCExperiment-method}
\alias{[,HiCExperiment-method}
\alias{[,HiCExperiment,numeric,ANY,ANY-method}
\alias{[,HiCExperiment,logical,ANY,ANY-method}
\alias{[,HiCExperiment,character,ANY,ANY-method}
\alias{seqinfo,HiCExperiment-method}
\alias{bins,HiCExperiment-method}
\alias{anchors,HiCExperiment-method}
\alias{regions,HiCExperiment-method}
\alias{show,HiCExperiment-method}
\alias{coerce,HiCExperiment,GInteractions-method}
\alias{coerce,HiCExperiment,ContactMatrix-method}
\alias{coerce,HiCExperiment,matrix-method}
\alias{coerce,HiCExperiment,data.frame-method}
\alias{as.matrix,HiCExperiment-method}
\alias{as.data.frame,HiCExperiment-method}
\title{\code{HiCExperiment} methods}
\arguments{
\item{x}{A \code{HiCExperiment} object.}

\item{object}{A \code{HiCExperiment} object.}

\item{name}{...}

\item{value}{...}

\item{i}{a range or boolean vector.}
}
\description{
HiCExperiment methods.
}
\examples{
library(HiCExperiment)
mcool_path <- HiContactsData::HiContactsData('yeast_wt', 'mcool')
contacts_yeast <- import(mcool_path, resolution = 16000, format = 'cool')
resolutions(contacts_yeast)
resolution(contacts_yeast)
focus(contacts_yeast)
scores(contacts_yeast)
tail(scores(contacts_yeast, 1))
tail(scores(contacts_yeast, 'balanced'))
scores(contacts_yeast, 'test') <- runif(length(contacts_yeast))
tail(scores(contacts_yeast, 'test'))
full_contacts_yeast <- contacts_yeast()
topologicalFeatures(full_contacts_yeast)
topologicalFeatures(full_contacts_yeast, 1)
topologicalFeatures(full_contacts_yeast, 'centromeres')
data(centros_yeast)
topologicalFeatures(contacts_yeast, 'centromeres') <- centros_yeast
topologicalFeatures(contacts_yeast, 'centromeres')
pairsFile(full_contacts_yeast)
fileName(contacts_yeast)
interactions(contacts_yeast)
length(contacts_yeast)
contacts_yeast[seq_len(10)]
seqinfo(contacts_yeast)
bins(contacts_yeast)
anchors(contacts_yeast)
regions(contacts_yeast)
contacts_yeast
as(contacts_yeast, 'GInteractions')
as(contacts_yeast, 'ContactMatrix')
as(contacts_yeast, 'matrix')[seq_len(10), seq_len(10)]
as(contacts_yeast, 'data.frame')[seq_len(10), seq_len(10)]
}
