% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{HiCExperiment_example}
\alias{HiCExperiment_example}
\alias{HiCExperimentCache}
\alias{.setHiCExperimentCache}
\title{Manage cache / download files for HiCExperiment package}
\usage{
HiCExperimentCache(directory = tools::R_user_dir("HiCExperiment", "cache"))

.setHiCExperimentCache(directory, ask = interactive())

HiCExperiment_example()
}
\arguments{
\item{directory}{character(1) The file location where the cache is located.
Once set, future downloads will go to this folder. See \code{setCache} section
for details.}

\item{ask}{logical(1) (default TRUE when \code{interactive()}) Confirm the file
location of the cache directory}
}
\value{
The directory / option of the cache location
}
\description{
Managing data downloads is important to save disk space and
re-downloading data files. This can be done effortlessly via the integrated
\code{BiocFileCache} system.
}
\section{HiCExperimentCache}{

Get the directory location of the cache. It will prompt the user to create
a cache if not already created. A specific directory can be used via
\code{setCache}.
}

\section{setCache}{

Specify the directory location of the data cache. By default, it will
go into the user's home and package name directory as given by
\link[tools:userdir]{R_user_dir} (default: varies by system e.g., for Linux:
'$HOME/.cache/R/HiCExperiment').
}

\examples{
getOption("HiCExperimentCache")
HiCExperimentCache()

}
