% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiCoolFile-class.R
\name{HiCoolFile-class}
\alias{HiCoolFile-class}
\alias{HiCoolFile}
\title{\code{HiCoolFile} S4 class}
\arguments{
\item{mcool}{String; path to a (m)cool file}

\item{resolution}{numeric; resolution to use with mcool file}

\item{pairs}{String; path to a pairs file}

\item{metadata}{list; contains path to log file and stats file from hicstuff}
}
\description{
The \code{HiCoolFile} class describes a \code{BiocFile} object, pointing to the location
of an (m)cool file and containing an additional \code{resolution} slot,
specifying at which resolution the associated mcool file should be parsed
The \code{HiCoolFile} class corresponds to an object linking several files
generated during the mapping and processing of paired-end sequencing reads
by \code{HiCool::HiCool()}. It extends the \code{CoolFile} data structure with
two additional slots:
\enumerate{
\item pairsFile: the path (in plain character) to the pairs file created
by \code{HiCool::HiCool()}, stored in a \code{PairsFile} object;
\item metadata: a list with two elements: \code{log} (path to \code{HiCool} processing
log file) and \code{stats} (aggregating some stats from \code{HiCool} mapping).
}
}
\section{Slots}{

\describe{
\item{\code{coolFile}}{CoolFile object}

\item{\code{pairsFile}}{PairsFile object}
}}

\examples{
mcool_path <- HiContactsData::HiContactsData('yeast_wt', 'mcool')
pairs_path <- HiContactsData::HiContactsData('yeast_wt', 'pairs.gz')
hcf <- HiCoolFile(mcool_path, resolution = 4000, pairs = pairs_path)
}
