% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiCoolFile-methods.R
\name{HiCoolFile-methods}
\alias{HiCoolFile-methods}
\alias{show,HiCoolFile-method}
\alias{pairsFile,HiCoolFile-method}
\title{\code{HiCoolFile} methods}
\arguments{
\item{object}{A \code{HiCoolFile} object.}

\item{x}{A \code{HiCoolFile} object.}
}
\description{
HiCoolFile methods.
}
\examples{
mcool_path <- HiContactsData::HiContactsData('yeast_wt', 'mcool')
pairs_path <- HiContactsData::HiContactsData('yeast_wt', 'pairs.gz')
hcf <- HiCoolFile(mcool_path, resolution = 4000, pairs = pairs_path)
hcf
resolution(hcf)
}
