% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HicproFile-class.R
\name{HicproFile-class}
\alias{HicproFile-class}
\alias{HicproFile}
\title{\code{HicproFile} S4 class}
\arguments{
\item{path}{String; path to the HiC-Pro output .matrix file (matrix file)}

\item{bed}{String; path to the HiC-Pro output .bed file (regions file)}

\item{pairsFile}{String; path to a pairs file}

\item{metadata}{list.}
}
\description{
The \code{HicproFile} class describes a \code{BiocFile} object, pointing to the location
of a .hic file and containing 3 additional slots:
\enumerate{
\item resolution: at which resolution the associated mcool file should be parsed
\item pairsFile: the path (in plain character) to an optional pairs file
(stored as a \code{PairsFile} object);
\item metadata: a list metadata
}
}
\section{Slots}{

\describe{
\item{\code{resolution}}{numeric value or NULL}

\item{\code{metadata}}{list}
}}

\examples{
hicpro_path <- HiContactsData::HiContactsData('yeast_wt', 'hic')
pairs_path <- HiContactsData::HiContactsData('yeast_wt', 'pairs.gz')
hicpro <- HicproFile(hicpro_path, pairsFile = pairs_path)
}
