% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HicproFile-methods.R
\name{HicproFile-methods}
\alias{HicproFile-methods}
\alias{show,HicproFile-method}
\title{\code{HicproFile} methods}
\arguments{
\item{object}{A \code{HicproFile} object.}
}
\description{
HicproFile methods.
}
\examples{
hicpro_matrix_path <- HiContactsData::HiContactsData('yeast_wt', 'hicpro_matrix')
hicpro_bed_path <- HiContactsData::HiContactsData('yeast_wt', 'hicpro_bed')
pairs_path <- HiContactsData::HiContactsData('yeast_wt', 'pairs.gz')
hcf <- HicproFile(hicpro_matrix_path, bed = hicpro_bed_path, pairs = pairs_path)
hcf
}
