% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-methods.R
\name{import-methods}
\alias{import-methods}
\alias{import}
\alias{import,CoolFile-method}
\alias{import,HiCoolFile-method}
\alias{import,PairsFile-method}
\title{HiCExperiment import methods}
\usage{
import(con, ...)
}
\arguments{
\item{con}{Path or connection to a cool, mcool or pairs file. Con argument
can also be a CoolFile or PairsFile object.}

\item{...}{e.g. \code{resolution = ...}; parameters to pass to the
format-specific method.}
}
\description{
Import methods for data structures implemented in the HiCExperiment package.
HiCExperiment package implements methods to faciliate
the import of (m)cool files and pairs files in R, as HiCExperiment
or GenomicInteractions objects.
}
