% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PairsFile-class.R, R/PairsFile-methods.R
\name{PairsFile-class}
\alias{PairsFile-class}
\alias{PairsFile}
\alias{pairsFile,PairsFile-method}
\title{\code{PairsFile} S4 class}
\arguments{
\item{x}{Path to a pairs file}
}
\description{
The \code{PairsFile} class describes a \code{BiocFile} object, pointing to the location
of pairs file, typically generated by \code{HiCool::HiCool()}.

PairsFile methods
}
\examples{
pairs_path <- HiContactsData::HiContactsData('yeast_wt', 'pairs.gz')
pf <- PairsFile(pairs_path)
pf
pairsFile(pf)
}
\seealso{
\code{\link[=CoolFile]{CoolFile()}}, \code{\link[=HiCFile]{HiCFile()}}, \code{\link[=HicProFile]{HicProFile()}}
}
