% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PairsFile-methods.R
\name{pairsFile-methods}
\alias{pairsFile-methods}
\alias{pairsFile,PairsFile-method}
\title{\code{pairsFile} methods}
\arguments{
\item{x}{Path to a pairs file}
}
\description{
PairsFile methods
}
\examples{
pairsFile(pf)
}
