% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PairsFile-class.R
\docType{class}
\name{PairsFile-class}
\alias{PairsFile-class}
\alias{PairsFile}
\title{\code{PairsFile} S4 class and methods}
\usage{
PairsFile(path)
}
\description{
\code{PairsFile} S4 class and methods
}
\examples{
pairs_path <- HiContactsData::HiContactsData('yeast_wt', 'pairs.gz')
PairsFile(pairs_path)
}
