% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-cool.R
\name{.getCoolAnchors}
\alias{.getCoolAnchors}
\alias{.getCountsFromPair}
\alias{.getCounts}
\alias{.fetchCool}
\alias{.dumpCool}
\alias{.lsCoolFiles}
\alias{lsCoolResolutions}
\alias{.cool2seqinfo}
\alias{.cool2gi}
\title{Parsing (m)cool files}
\usage{
.getCoolAnchors(file, resolution = NULL, balanced = "cooler")

.getCountsFromPair(file, pair, anchors, resolution = NULL)

.getCounts(file, coords, anchors, resolution = NULL)

.fetchCool(file, path, resolution = NULL, idx = NULL, ...)

.dumpCool(file, resolution = NULL)

.lsCoolFiles(file, verbose = FALSE)

lsCoolResolutions(file, verbose = FALSE)

.cool2seqinfo(file, resolution = NULL)

.cool2gi(file, coords = NULL, resolution = NULL)
}
\arguments{
\item{file}{file}

\item{resolution}{resolution}

\item{balanced}{import balancing scores}

\item{pair}{pair
(e.g. S4Vectors::Pairs(GRanges("II:200000-300000"), GRanges("II:70000-100000"))).}

\item{anchors}{anchors}

\item{coords}{NULL, character, or GRanges.
Can also be a Pairs object of paired GRanges (length of 1).}

\item{path}{path}

\item{idx}{idx to extract in cool file}

\item{...}{...}

\item{verbose}{Print resolutions in the console}
}
\value{
anchors from (m)cool, stored as a GRanges

counts from (m)cool, stored as a tibble

counts from (m)cool, stored as a tibble

vector

vector

vector

a Seqinfo object

a GenomicInteractions object
}
\description{
These functions are the workhorse internal functions used to import
a \verb{.(m)cool} file as GenomicInteractions (wrapped into a \code{HiCExperiment} object
by \code{HiCExperiment()} function).
}
\examples{
mcool_path <- HiContactsData::HiContactsData('yeast_wt', 'mcool')
lsCoolResolutions(mcool_path, verbose = TRUE)
}
