% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-other.R
\name{gi2cm}
\alias{gi2cm}
\alias{cm2matrix}
\title{Other parsing functions}
\usage{
gi2cm(gi)

cm2matrix(cm, replace_NA = NA)
}
\arguments{
\item{gi}{GInteractions}

\item{cm}{A \code{ContactMatrix} object}

\item{replace_NA}{Replace NA values}
}
\value{
a dense matrix
}
\description{
Other parsing functions
}
\examples{
mcool_path <- HiContactsData::HiContactsData('yeast_wt', 'mcool')
contacts <- import(mcool_path, format = 'cool')
gis <- interactions(contacts)
gis$score <- scores(contacts, 1)
cm <- gi2cm(gis)
cm
cm2matrix(cm)[1:10, 1:10]
}
