% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{splitCoords}
\alias{splitCoords}
\alias{coords2char}
\alias{char2coords}
\alias{fullContactInteractions}
\alias{sortPairs}
\alias{asGInteractions}
\title{Utils functions}
\usage{
splitCoords(coords)

coords2char(coords, big.mark = ",")

char2coords(char)

fullContactInteractions(chr, start, end, binning)

sortPairs(pairs)

asGInteractions(df)
}
\arguments{
\item{coords}{coords}

\item{big.mark}{big.mark}

\item{char}{char (e.g. "II:30000-50000" or "II:30000-50000 x II:60000-80000")}

\item{chr}{chr}

\item{start}{start}

\item{end}{end}

\item{binning}{binning}

\item{pairs}{pairs}

\item{df}{df}
}
\value{
a list containing \code{chr}, \code{start} and \code{end}

a character string

a S4Vectors::Pairs object

a GenomicInteractions object

a Pairs object

a GenomicInteractions object
}
\description{
Utilities to facilitate parsing/handling of coordinates, GInteractions,
Pairs, ...
}
