% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggrHiCExperiment-class.R
\docType{class}
\name{AggrHiCExperiment}
\alias{AggrHiCExperiment}
\alias{.aggrHiCExperiment}
\title{\code{AggrHiCExperiment} S4 class}
\usage{
AggrHiCExperiment(
  file,
  resolution = NULL,
  targets,
  metadata = list(),
  topologicalFeatures = S4Vectors::SimpleList(),
  pairsFile = NULL,
  bed = NULL,
  BPPARAM = BiocParallel::bpparam()
)

.aggrHiCExperiment(
  file,
  resolution = NULL,
  targets,
  metadata = list(),
  topologicalFeatures = S4Vectors::SimpleList(),
  pairsFile = NULL,
  bed = NULL,
  BPPARAM
)
}
\arguments{
\item{file}{CoolFile or plain path to a Hi-C contact file}

\item{resolution}{Resolution to use with the Hi-C contact file}

\item{targets}{Set of chromosome coordinates for which
interaction counts are extracted from the Hi-C contact file, provided
as a GRanges object (for diagnoal-centered loci) or as a GInteractions
object (for off-diagonal coordinates).}

\item{metadata}{list of metadata}

\item{topologicalFeatures}{topologicalFeatures provided as a named SimpleList}

\item{pairsFile}{Path to an associated .pairs file}

\item{bed}{Path to regions file generated by HiC-Pro}

\item{BPPARAM}{BiocParallel parameters}
}
\description{
The \code{AggrHiCExperiment} extends \code{HiCExperiment} class
}
\section{Slots}{

\describe{
\item{\code{fileName}}{Path of Hi-C contact file}

\item{\code{resolutions}}{Resolutions available in the Hi-C contact file.}

\item{\code{resolution}}{Current resolution}

\item{\code{interactions}}{Genomic Interactions extracted from the Hi-C contact file}

\item{\code{scores}}{Available interaction scores.}

\item{\code{slices}}{Available interaction slices.}

\item{\code{topologicalFeatures}}{Topological features associated with the dataset
(e.g. loops (\<Pairs\>), borders (\<GRanges\>),
viewpoints (\<GRanges\>), etc...)}

\item{\code{pairsFile}}{Path to the .pairs file associated with the Hi-C contact file}

\item{\code{metadata}}{metadata associated with the Hi-C contact file.}
}}

\examples{
microC <- HiCExperiment_example()
CTCF_peaks <- topologicalFeatures(microC, 'CTCF_peaks')
x <- AggrHiCExperiment(
  file = fileName(microC), 
  resolution = resolution(microC),
  targets = CTCF_peaks
)
x
slices(x, 'count')[, , 1]
}
