% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggrHiCExperiment-class.R,
%   R/AggrHiCExperiment-methods.R
\docType{class}
\name{AggrHiCExperiment}
\alias{AggrHiCExperiment}
\alias{AggrHiCExperiment-class}
\alias{.aggrHiCExperiment}
\title{\code{AggrHiCExperiment} S4 class}
\usage{
AggrHiCExperiment(
  file,
  resolution = NULL,
  snippets,
  metadata = list(),
  topologicalFeatures = S4Vectors::SimpleList(),
  pairsFile = NULL,
  BPPARAM = BiocParallel::bpparam(),
  ...
)

.aggrHiCExperiment(
  file,
  resolution = NULL,
  snippets,
  metadata = list(),
  topologicalFeatures = S4Vectors::SimpleList(),
  pairsFile = NULL,
  BPPARAM
)
}
\description{
The \code{AggrHiCExperiment} extends \code{HiCExperiment} class

The \code{AggrHiCExperiment} extends \code{HiCExperiment} class
}
\section{Slots}{

\describe{
\item{\code{fileName}}{Path of Hi-C contact file}

\item{\code{resolutions}}{Resolutions available in the Hi-C contact file.}

\item{\code{resolution}}{Current resolution}

\item{\code{interactions}}{Genomic Interactions extracted from the Hi-C contact file}

\item{\code{scores}}{Available interaction scores.}

\item{\code{topologicalFeatures}}{Topological features associated with the dataset
(e.g. loops (\<Pairs\>), borders (\<GRanges\>),
viewpoints (\<GRanges\>), etc...)}

\item{\code{pairsFile}}{Path to the .pairs file associated with the Hi-C contact file}

\item{\code{metadata}}{metadata associated with the Hi-C contact file.}

\item{\code{fileName}}{Path of Hi-C contact file}

\item{\code{resolutions}}{Resolutions available in the Hi-C contact file.}

\item{\code{resolution}}{Current resolution}

\item{\code{interactions}}{Genomic Interactions extracted from the Hi-C contact file}

\item{\code{scores}}{Available interaction scores.}

\item{\code{topologicalFeatures}}{Topological features associated with the dataset
(e.g. loops (\<Pairs\>), borders (\<GRanges\>),
viewpoints (\<GRanges\>), etc...)}

\item{\code{pairsFile}}{Path to the .pairs file associated with the Hi-C contact file}

\item{\code{metadata}}{metadata associated with the Hi-C contact file.}
}}

\examples{
library(rtracklayer)
devtools::load_all('.')
file = '/home/rsg/.cache/R/ExperimentHub/7fa45373d163_7836'
file = "/home/rsg/.cache/R/fourDNData/78685597e138_4DNFI9FVHJZQ.mcool"
file = "/data/20221214_HiContacts_compartments-TADs-loops/data/test.mcool"
resolution = 10000 
BPPARAM = BiocParallel::SerialParam(progressbar = TRUE)
coords_list = import('/home/rsg/Projects/20221214_HiContacts_compartments-TADs-loops/data/test_CTCF.bed.gz')
CTCF_bw <- import('/home/rsg/Projects/20221214_HiContacts_compartments-TADs-loops/data/test_CTCF.bigWig', as = 'NumericList', selection = BigWigSelection(resize(coords_list, fix = 'center', width = 250)))
coords_list = import('/home/rsg/Projects/20221214_HiContacts_compartments-TADs-loops/data/test_CTCF.bed.gz')
coords_list$scoreChIP <- mean(CTCF_bw)
snippets <- coords_list[coords_list$score >= quantile(coords_list$score , 0.75) & coords_list$scoreChIP >= quantile(coords_list$scoreChIP , 0.75)]
snippets <- GenomicRanges::resize(snippets, width = resolution*100, fix = 'center')
snippets <- snippets[order(snippets$scoreChIP, decreasing = TRUE)]
#snippets <- snippets[strand(snippets) == '+']
#snippets <- snippets[resize(snippets, fix = 'center', width = 1) \%over\% GRanges('chr2:1-12000000')]
pairs <- GInteractions(
    rep(snippets, each = length(snippets)),
    rep(snippets, length(snippets))
)
pairs <- swapAnchors(pairs)
pairs <- pairs[!is.na(pairdist(pairs))]
pairs <- pairs[pairdist(pairs) > 30000 & pairdist(pairs) < 500000]
}
