% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{AllGenerics}
\alias{AllGenerics}
\alias{availableResolutions}
\alias{availableChromosomes}
\alias{resolutions}
\alias{resolution}
\alias{focus}
\alias{focus<-}
\alias{scores}
\alias{scores<-}
\alias{topologicalFeatures}
\alias{topologicalFeatures<-}
\alias{pairsFile}
\alias{pairsFile<-}
\alias{metadata<-}
\alias{bins}
\alias{slices}
\alias{zoom}
\alias{refocus}
\alias{cis}
\alias{trans}
\alias{bin}
\title{Generic functions}
\arguments{
\item{x}{Passed to corresponding method}

\item{name}{Passed to corresponding method}

\item{value}{Passed to corresponding method}

\item{...}{Passed to corresponding method}
}
\description{
Generics functions created in HiCExperiment package.
}
\keyword{internal}
