% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin-methods}
\alias{bin-methods}
\alias{bin,GInteractions,character-method}
\alias{bin,GInteractions,numeric-method}
\alias{bin,PairsFile,numeric-method}
\title{HiCExperiment binning methods}
\usage{
\S4method{bin}{GInteractions,numeric}(x, resolution, seqinfo = NULL)

\S4method{bin}{PairsFile,numeric}(x, resolution, seqinfo = NULL)
}
\arguments{
\item{x}{A PairsFile or GInteractions object}

\item{resolution}{Which resolution to use to bin the interactions}

\item{seqinfo}{Seqinfo object}
}
\description{
HiCExperiment binning methods
}
\examples{
pairsf <- HiContactsData::HiContactsData('yeast_wt', 'pairs.gz')
pf <- PairsFile(pairsf)
}
