% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoolFile-methods.R
\name{CoolFile-methods}
\alias{CoolFile-methods}
\alias{show,CoolFile-method}
\alias{pairsFile,CoolFile-method}
\alias{resolution,CoolFile-method}
\alias{metadata<-,CoolFile-method}
\alias{metadata<-,CoolFile,list-method}
\title{\code{CoolFile} methods}
\arguments{
\item{object}{A \code{CoolFile} object.}

\item{x}{A \code{CoolFile} object.}
}
\description{
CoolFile methods.
}
\examples{
mcool_path <- HiContactsData::HiContactsData('yeast_wt', 'mcool')
pairs_path <- HiContactsData::HiContactsData('yeast_wt', 'pairs.gz')
hcf <- CoolFile(mcool_path, resolution = 4000, pairs = pairs_path)
hcf
resolution(hcf)
}
