% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoolFile-methods.R
\name{CoolFile-methods}
\alias{CoolFile-methods}
\alias{show,CoolFile-method}
\title{\code{CoolFile} methods}
\arguments{
\item{object}{A \code{CoolFile} object.}
}
\description{
CoolFile methods.
}
\examples{
mcool_path <- HiContactsData::HiContactsData('yeast_wt', 'mcool')
pairs_path <- HiContactsData::HiContactsData('yeast_wt', 'pairs.gz')
cf <- CoolFile(mcool_path, resolution = 4000, pairsFile = pairs_path, metadata = list(info = 'Yeast WT Hi-C exp.'))
cf
resolution(cf)
pairsFile(cf)
metadata(cf)
}
