% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Example datasets provided in \code{HiCExperiment} & \code{HiContactsData}}
\format{
An object of class \code{"GRanges"}.
}
\source{
HiContacts
}
\usage{
data(centros_yeast)
}
\description{
Example datasets provided in \code{HiCExperiment} & \code{HiContactsData}
}
\examples{
data(centros_yeast)
centros_yeast
contacts_yeast()
contacts_yeast_eco1()
full_contacts_yeast()
}
