% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-methods.R
\name{export-methods}
\alias{export-methods}
\alias{export}
\alias{export,HiCExperiment,character,missing-method}
\title{HiCExperiment export methods}
\usage{
export(object, con)
}
\arguments{
\item{object}{HiCExperiment object}

\item{con}{Prefix of \verb{*_regions.bed} and \verb{*_matrix.mtx} files}
}
\value{
A \code{HicproFile} object
}
\description{
Export methods to save \code{HiCExperiment} objects into HiC-Pro-like \code{regions}
and \code{matrix} files (\code{regions.bed} and \code{matrix.mtx})
}
\examples{
mcoolPath <- HiContactsData::HiContactsData('yeast_wt', 'mcool')
x <- import(mcoolPath, resolution = 16000, focus = 'XVI', format = 'cool')
}
