% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiCExperiment-class.R, R/HiCExperiment-method.R
\docType{class}
\name{HiCExperiment-class}
\alias{HiCExperiment-class}
\alias{HiCExperiment}
\alias{resolutions}
\alias{resolutions,HiCExperiment-method}
\alias{resolution}
\alias{resolution,HiCExperiment-method}
\alias{focus}
\alias{focus,HiCExperiment-method}
\alias{focus<-}
\alias{focus<-,HiCExperiment,character-method}
\alias{scores}
\alias{scores,HiCExperiment,missing-method}
\alias{scores,HiCExperiment,character-method}
\alias{scores,HiCExperiment,numeric-method}
\alias{scores<-}
\alias{scores<-,HiCExperiment,character,numeric-method}
\alias{topologicalFeatures}
\alias{topologicalFeatures,HiCExperiment,missing-method}
\alias{topologicalFeatures,HiCExperiment,character-method}
\alias{topologicalFeatures,HiCExperiment,numeric-method}
\alias{topologicalFeatures<-}
\alias{topologicalFeatures<-,HiCExperiment,character,GRangesOrGInteractions-method}
\alias{pairsFile}
\alias{pairsFile,HiCExperiment-method}
\alias{pairsFile<-}
\alias{pairsFile<-,HiCExperiment,character-method}
\alias{metadata<-}
\alias{metadata<-,HiCExperiment,list-method}
\alias{fileName}
\alias{fileName,HiCExperiment-method}
\alias{interactions}
\alias{interactions,HiCExperiment-method}
\alias{interactions<-}
\alias{interactions<-,HiCExperiment,GInteractions-method}
\alias{length}
\alias{length,HiCExperiment-method}
\alias{[}
\alias{[,HiCExperiment,numeric,ANY,ANY-method}
\alias{[,HiCExperiment,logical,ANY,ANY-method}
\alias{[,HiCExperiment,character,ANY,ANY-method}
\alias{seqinfo,HiCExperiment-method}
\alias{bins}
\alias{bins,HiCExperiment-method}
\alias{anchors}
\alias{anchors,HiCExperiment-method}
\alias{regions}
\alias{regions,HiCExperiment-method}
\alias{summary}
\alias{summary,HiCExperiment-method}
\alias{show}
\alias{show,HiCExperiment-method}
\alias{setAs}
\alias{setAs,HiCExperiment-method}
\title{\code{HiCExperiment} S4 class and methods}
\usage{
HiCExperiment(
  file,
  resolution = NULL,
  focus = NULL,
  metadata = list(),
  topologicalFeatures = S4Vectors::SimpleList(loops =
    S4Vectors::Pairs(GenomicRanges::GRanges(), GenomicRanges::GRanges()), borders =
    GenomicRanges::GRanges(), compartments = GenomicRanges::GRanges(), viewpoints =
    GenomicRanges::GRanges()),
  pairsFile = NULL
)

\S4method{resolutions}{HiCExperiment}(x)

\S4method{resolution}{HiCExperiment}(x)

\S4method{focus}{HiCExperiment}(x)

\S4method{focus}{HiCExperiment,character}(x) <- value

\S4method{scores}{HiCExperiment,missing}(x)

\S4method{scores}{HiCExperiment,character,numeric}(x, name) <- value

\S4method{topologicalFeatures}{HiCExperiment,missing}(x)

\S4method{topologicalFeatures}{HiCExperiment,character,GRangesOrGInteractions}(x, name) <- value

\S4method{pairsFile}{HiCExperiment}(x)

\S4method{pairsFile}{HiCExperiment,character}(x) <- value

\S4method{metadata}{HiCExperiment,list}(x) <- value

\S4method{fileName}{HiCExperiment}(object)

\S4method{interactions}{HiCExperiment}(x)

\S4method{interactions}{HiCExperiment,GInteractions}(x) <- value

\S4method{length}{HiCExperiment}(x)

\S4method{[}{HiCExperiment,numeric,ANY,ANY}(x, i)

\S4method{seqinfo}{HiCExperiment}(x)

\S4method{bins}{HiCExperiment}(x)

\S4method{anchors}{HiCExperiment}(x)

\S4method{regions}{HiCExperiment}(x)

\S4method{summary}{HiCExperiment}(object)

\S4method{show}{HiCExperiment}(object)
}
\arguments{
\item{file}{Path to a (m)cool file}

\item{resolution}{Resolution to use with mcool file}

\item{focus}{focus Chr. coordinates for which
interaction counts are extracted from the .(m)cool file, provided
as a character string (e.g. "II:4000-5000"). If not provided,
the entire (m)cool file will be imported.}

\item{metadata}{list of metadata}

\item{topologicalFeatures}{topologicalFeatures provided as a named SimpleList}

\item{pairsFile}{Path to an associated .pairs file}

\item{x}{A \code{HiCExperiment} object.}

\item{value}{value}

\item{name}{name}

\item{object}{A \code{HiCExperiment} object.}

\item{i}{a range or boolean vector.}
}
\value{
a new \code{HiCExperiment} object.
}
\description{
\code{HiCExperiment} S4 class and methods
}
\section{Slots}{

\describe{
\item{\code{fileName}}{Path of (m)cool file}

\item{\code{focus}}{Chr. coordinates for which interaction counts are extracted
from the .(m)cool file.}

\item{\code{resolutions}}{Resolutions available in the .(m)cool file.}

\item{\code{resolution}}{Current resolution}

\item{\code{interactions}}{Genomic Interactions extracted from the .(m)cool object}

\item{\code{scores}}{Available interaction scores.}

\item{\code{topologicalFeatures}}{Topological features associated with the dataset
(e.g. loops (\<Pairs\>), borders (\<GRanges\>),
viewpoints (\<GRanges\>), etc...)}

\item{\code{pairsFile}}{Path to the .pairs file associated with the .(m)cool file}

\item{\code{metadata}}{metadata associated with the .(m)cool file.}
}}

\examples{
library(HiCExperiment)
contacts_yeast <- contacts_yeast()
contacts_yeast
library(HiCExperiment)
contacts_yeast <- contacts_yeast()
resolutions(contacts_yeast)
resolution(contacts_yeast)
focus(contacts_yeast)
scores(contacts_yeast)
tail(scores(contacts_yeast, 1))
tail(scores(contacts_yeast, 'balanced'))
scores(contacts_yeast, 'test') <- runif(length(contacts_yeast))
tail(scores(contacts_yeast, 'test'))
full_contacts_yeast <- full_contacts_yeast()
topologicalFeatures(full_contacts_yeast)
topologicalFeatures(full_contacts_yeast, 1)
topologicalFeatures(full_contacts_yeast, 'centromeres')
data(centros_yeast)
topologicalFeatures(contacts_yeast, 'centromeres') <- centros_yeast
topologicalFeatures(contacts_yeast, 'centromeres')
pairsFile(full_contacts_yeast)
fileName(contacts_yeast)
interactions(contacts_yeast)
length(contacts_yeast)
contacts_yeast[seq_len(10)]
seqinfo(contacts_yeast)
bins(contacts_yeast)
anchors(contacts_yeast)
regions(contacts_yeast)
summary(contacts_yeast)
show(contacts_yeast)
as(contacts_yeast, 'GInteractions')
as(contacts_yeast, 'ContactMatrix')
as(contacts_yeast, 'matrix')[seq_len(10), seq_len(10)]
as(contacts_yeast, 'data.frame')[seq_len(10), seq_len(10)]
}
