% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiCExperiment-class.R
\docType{class}
\name{HiCExperiment}
\alias{HiCExperiment}
\alias{HiCExperiment-class}
\alias{.HiCExperimentFromCoolFile}
\alias{.HiCExperimentFromHicFile}
\alias{.HiCExperimentFromHicproFile}
\title{\code{HiCExperiment} S4 class}
\usage{
HiCExperiment(
  file,
  resolution = NULL,
  focus = NULL,
  metadata = list(),
  topologicalFeatures = S4Vectors::SimpleList(loops =
    S4Vectors::Pairs(GenomicRanges::GRanges(), GenomicRanges::GRanges()), borders =
    GenomicRanges::GRanges(), compartments = GenomicRanges::GRanges(), viewpoints =
    GenomicRanges::GRanges()),
  pairsFile = NULL,
  ...
)

.HiCExperimentFromCoolFile(
  file,
  resolution = NULL,
  focus = NULL,
  metadata = list(),
  topologicalFeatures = S4Vectors::SimpleList(loops =
    S4Vectors::Pairs(GenomicRanges::GRanges(), GenomicRanges::GRanges()), borders =
    GenomicRanges::GRanges(), compartments = GenomicRanges::GRanges(), viewpoints =
    GenomicRanges::GRanges()),
  pairsFile = NULL
)

.HiCExperimentFromHicFile(
  file,
  resolution = NULL,
  focus = NULL,
  metadata = list(),
  topologicalFeatures = S4Vectors::SimpleList(loops =
    S4Vectors::Pairs(GenomicRanges::GRanges(), GenomicRanges::GRanges()), borders =
    GenomicRanges::GRanges(), compartments = GenomicRanges::GRanges(), viewpoints =
    GenomicRanges::GRanges()),
  pairsFile = NULL
)

.HiCExperimentFromHicproFile(
  file,
  bed,
  metadata = list(),
  topologicalFeatures = S4Vectors::SimpleList(loops =
    S4Vectors::Pairs(GenomicRanges::GRanges(), GenomicRanges::GRanges()), borders =
    GenomicRanges::GRanges(), compartments = GenomicRanges::GRanges(), viewpoints =
    GenomicRanges::GRanges()),
  pairsFile = NULL
)
}
\arguments{
\item{file}{CoolFile or plain path to a Hi-C contact file}

\item{resolution}{Resolution to use with the Hi-C contact file}

\item{focus}{Chromosome coordinates for which
interaction counts are extracted from the Hi-C contact file, provided
as a character string (e.g. "II:4000-5000"). If not provided,
the entire Hi-C contact file will be imported.}

\item{metadata}{list of metadata}

\item{topologicalFeatures}{topologicalFeatures provided as a named SimpleList}

\item{pairsFile}{Path to an associated .pairs file}

\item{...}{Extra arguments}

\item{bed}{Path to regions file generated by HiC-Pro}
}
\value{
An \code{HiCExperiment} object.
}
\description{
The \code{HiCExperiment} class describes Hi-C contact files imported in R, either
through the \code{HiCExperiment} constructor function or using the \code{import}
method implemented by \code{HiCExperiment} package.
}
\section{Slots}{

\describe{
\item{\code{fileName}}{Path of Hi-C contact file}

\item{\code{focus}}{Chr. coordinates for which interaction counts are extracted
from the Hi-C contact file.}

\item{\code{resolutions}}{Resolutions available in the Hi-C contact file.}

\item{\code{resolution}}{Current resolution}

\item{\code{interactions}}{Genomic Interactions extracted from the Hi-C contact file}

\item{\code{scores}}{Available interaction scores.}

\item{\code{topologicalFeatures}}{Topological features associated with the dataset
(e.g. loops (\<Pairs\>), borders (\<GRanges\>),
viewpoints (\<GRanges\>), etc...)}

\item{\code{pairsFile}}{Path to the .pairs file associated with the Hi-C contact file}

\item{\code{metadata}}{metadata associated with the Hi-C contact file.}
}}

\examples{
library(HiCExperiment)
mcool_path <- HiContactsData::HiContactsData('yeast_wt', 'mcool')
HiCExperiment(mcool_path, resolution = 16000)
}
