% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HicFile-methods.R
\name{HicFile-methods}
\alias{HicFile-methods}
\alias{show,HicFile-method}
\title{\code{HicFile} methods}
\arguments{
\item{object}{A \code{HicFile} object.}
}
\description{
HicFile methods.
}
\examples{
hic_path <- HiContactsData::HiContactsData('yeast_wt', 'hic')
pairs_path <- HiContactsData::HiContactsData('yeast_wt', 'pairs.gz')
hcf <- HicFile(hic_path, resolution = 4000, pairsFile = pairs_path)
hcf
}
