% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-slices.R
\name{multi2Query}
\alias{multi2Query}
\alias{.multi2DQuery}
\title{Querying multiple slices of a contact matrix}
\usage{
.multi2DQuery(
  file,
  resolution,
  pairs,
  max.distance = NULL,
  bed = NULL,
  BPPARAM = BiocParallel::bpparam()
)
}
\arguments{
\item{file}{file}

\item{resolution}{resolution}

\item{pairs}{slices to read}

\item{max.distance}{Maximum distance to use when compiling distance decay}

\item{bed}{associated bed file for HiC-Pro derived contact matrix.}

\item{BPPARAM}{BiocParallel parameters}
}
\value{
a GInteractions object with \code{count}, \code{balanced}, \code{detrended} and
\code{expected} scores
}
\description{
These functions are the workhorse internal functions used to extract
counts from multiple genomic coordinates in a Hi-C contact matrix.
}
