% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-cool.R
\name{.getAnchors}
\alias{.getAnchors}
\title{Parsing (m)cool files}
\usage{
.getAnchors(file, resolution = NULL, balanced = "cooler")
}
\arguments{
\item{file}{file}

\item{resolution}{resolution}

\item{balanced}{import balancing scores}
}
\value{
anchors from (m)cool, stored as a GRanges
}
\description{
These functions are the workhorse internal functions used to import
a \verb{.(m)cool} file as GenomicInteractions (wrapped into a \code{HiCExperiment} object
by \code{HiCExperiment()} function).
}
