% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-hic.R
\name{.hic2gi}
\alias{.hic2gi}
\alias{lsHicResolutions}
\alias{.getHicAnchors}
\alias{.hic2seqinfo}
\title{Parsing hic files}
\usage{
.hic2gi(file, coords = NULL, resolution = NULL)

lsHicResolutions(file, verbose = FALSE)

.getHicAnchors(file, resolution = NULL)

.hic2seqinfo(file)
}
\arguments{
\item{file}{file}

\item{coords}{NULL, character, or GRanges.
Can also be a Pairs object of paired GRanges (length of 1).}

\item{resolution}{resolution}

\item{verbose}{Print resolutions in the console}
}
\value{
a GenomicInteractions object

vector
}
\description{
These functions are the workhorse internal functions used to import
a \code{.hic} file as GenomicInteractions (wrapped into a \code{HiCExperiment} object
by \code{HiCExperiment()} function).
}
\examples{
hic_path <- HiContactsData::HiContactsData('yeast_wt', 'hic')
lsHicResolutions(hic_path, verbose = TRUE)
}
