% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-hic.R
\name{.hic2gi}
\alias{.hic2gi}
\title{Parsing hic files}
\usage{
.hic2gi(file, coords = NULL, resolution = NULL)
}
\arguments{
\item{file}{file}

\item{coords}{NULL, character, or GRanges.
Can also be a Pairs object of paired GRanges (length of 1).}

\item{resolution}{resolution}
}
\value{
a GenomicInteractions object
}
\description{
These functions are the workhorse internal functions used to import
a \code{.hic} file as GenomicInteractions (wrapped into a \code{HiCExperiment} object
by \code{HiCExperimentFromHicFile()} function).
}
