% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-hicpro.R
\name{.hicpro2gi}
\alias{.hicpro2gi}
\alias{.getHicproAnchors}
\alias{.hicpro2seqinfo}
\alias{.dumpHicpro}
\title{Parsing hicpro files (matrix & bed)}
\usage{
.hicpro2gi(file, bed)

.getHicproAnchors(bed)

.hicpro2seqinfo(bed)

.dumpHicpro(file, bed)
}
\arguments{
\item{file}{path to a matrix file from HiC-Pro}

\item{bed}{path to the regions file generated by HiC-Pro}
}
\value{
a GInteractions object
}
\description{
These functions are the workhorse internal functions used to import
HiC-Pro \code{.matrix} and \code{.bed} files as GInteractions (wrapped into a \code{HiCExperiment} object
by \code{HiCExperiment()} function).
}
