% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-other.R
\name{gi2cm}
\alias{gi2cm}
\alias{cm2matrix}
\title{Other parsing functions}
\usage{
gi2cm(gi, use.scores = "score")

cm2matrix(cm, replace_NA = NA)
}
\arguments{
\item{gi}{GInteractions}

\item{use.scores}{Which scores to use to inflate GInteractions}

\item{cm}{A \code{ContactMatrix} object}

\item{replace_NA}{Replace NA values}
}
\value{
a dense matrix
}
\description{
Other parsing functions
}
\examples{
mcool_path <- HiContactsData::HiContactsData('yeast_wt', 'mcool')
contacts <- import(mcool_path, focus = 'XVI', resolution = 16000, format = 'cool')
gis <- interactions(contacts)
cm <- gi2cm(gis, 'balanced')
cm
cm2matrix(cm)[1:10, 1:10]
}
