% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{HiCExperiment utils}
\alias{HiCExperiment utils}
\alias{splitCoords}
\alias{coords2char}
\alias{char2coords}
\alias{sortPairs}
\alias{asGInteractions}
\title{Utils functions}
\usage{
splitCoords(coords)

coords2char(coords, big.mark = ",")

char2coords(char)

sortPairs(pairs)

asGInteractions(df)
}
\arguments{
\item{coords}{coords}

\item{big.mark}{big.mark}

\item{char}{char (e.g. "II:30000-50000" or "II:30000-50000|II:60000-80000")}

\item{pairs}{pairs}

\item{df}{df}
}
\value{
Reformatted coordinates or GInteractions
}
\description{
Utilities to facilitate parsing/handling of coordinates, GInteractions,
Pairs, ...
}
